Eager load ingredients when fetching a recipe
Most times a recipe is shown the ingredient additions of fermentables,
hops, and yeasts will also be fetched. In the average recipe with say 4
fermentables, 3 hop additions, a 1 yeast, this can start adding up to a
few additional queries that could be handled more effectively by eager
loading when the recipe is fetched.

Putting this logic in the controller is likely not ideal and I should go
back and put it in the model as something like
Recipe#find_with_ingredients to wrap it up.