Move BeerXml module back to app/models dir
I had pulled this out to a plugin long long ago. It wasn't really ready
for that so I moved it to the lib/ dir when updating to Rails 3. At this
point there is no benefit of keeping it separate and it is better to
keep all the model logic together until there is more reason to pull it
out.