Add model for the Tinseth hop utilization formula
This formula is used to calculate the amount of alpha acid that is
likely to be extracted from a hop addition. This varies based on the
amount of time the hops are boiled and the gravity of the wort being
boiled also effects the rate of isomerization.

This utilization factor is most important to calculate IBUs, or
international bitterness units, for a recipe.