Pre-populate hop addition form with hop defaults
When selecting a type of hop to add to a recipe, it is convenient to
have the alpha acid and beta acid values filled out with the average
values for that type of hop. These values vary from each harvest of hops
but it's a good starting point for planning recipes if you don't have
the actual hops on hand to know the exact values you will be using.