Split out recipe profile attributes to new model
There will be more calculated attributes added for a Recipe so it makes
sense to move these over to a RecipeProfile model now. This mimics the
current UI but also makes sense on the object level if the UI were to
change away from this representation. This way we keep the Recipe model
more focused on the responsibility of being an ActiveRecord persistance object